fix(test): extract _collectDecompressed to shared helper, wire test into test:sidecar
- Extract _collectDecompressed from ais-relay.cjs into scripts/_relay-decompress.cjs
  so tests can import the real function without triggering relay startup/process.exit
- Update ais-relay-gzip.test.cjs to require the real implementation instead of
  duplicating it (eliminates drift risk between test copy and production code)
- Add scripts/ais-relay-gzip.test.cjs to test:sidecar in package.json so the
  maxBytes regression test runs in pre-push and CI automatically
