+++ scripts/ais-relay.cjs
@@ -5586,28 +5586,7 @@ async function _fetchOpenSkyToken(clientId, clientSecret) {
 }
 
 // Promisified upstream OpenSky fetch (single request)
-function _collectDecompressed(response, maxBytes) {
-  return new Promise((resolve, reject) => {
-    const enc = (response.headers['content-encoding'] || '').trim().toLowerCase();
-    let stream = response;
-    if (enc === 'gzip' || enc === 'x-gzip') stream = response.pipe(zlib.createGunzip());
-    else if (enc === 'deflate') stream = response.pipe(zlib.createInflate());
-    else if (enc === 'br') stream = response.pipe(zlib.createBrotliDecompress());
-    const chunks = [];
-    let totalSize = 0;
-    stream.on('data', chunk => {
-      totalSize += chunk.length;
-      if (maxBytes && totalSize > maxBytes) {
-        stream.destroy();
-        response.destroy();
-        return reject(new Error(`payload exceeds ${maxBytes} byte limit (${totalSize} bytes decompressed)`));
-      }
-      chunks.push(chunk);
-    });
-    stream.on('end', () => resolve(Buffer.concat(chunks).toString()));
-    stream.on('error', (err) => reject(new Error(`decompression failed (${enc}): ${err.message}`)));
-  });
-}
+const { _collectDecompressed } = require('./_relay-decompress.cjs');
 
 function _openskyRawFetch(url, token) {
   const parsed = new URL(url);
